food-microservice/src/main/java/pizzeria/food/domain/ingredient/InvalidIngredientException.java
package pizzeria.food.domain.ingredient;

public class InvalidIngredientException extends Exception {
    static final long serialVersionUID = -3437516993124229568L;

    /**
     * Constructs a new exception with {@code null} as its detail message.
     * The cause is not initialized, and may subsequently be initialized by a
     * call to {@link #initCause}.
     */
    public InvalidIngredientException() {
        super("The ingredient's properties are invalid");
    }

    /**
     * Returns the detail message string of this throwable.
     *
     * @return the detail message string of this {@code Throwable} instance
     * (which may be {@code null}).
     */
    @Override
    public String getMessage() {
        return super.getMessage();
    }
}
